requirements.txt
openai
python-dotenv
aiohttp
streamlit
flask